Fix new views not properly inheriting correct text size
float value for the text size multiplier was being cast to long long
(NSInteger) invalidating the decimal position

+++ Classes/Views/Channel View/TVCLogController.m
@@ -1514,7 +1514,7 @@
 	[[self printingQueue] updateReadinessState:self];
 
 	/* Change the font size to the one of others for new views. */
-	NSInteger math = [worldController() textSizeMultiplier];
+	float math = [worldController() textSizeMultiplier];
 
 	[self.webView setTextSizeMultiplier:math];
 }
